fix(auth): SSR session guard with @supabase/ssr; oauth callback respects redirect; precheck email before OTP; forgot-password redirects to /login; prevent OTP on Google sign-in
